Revert "drm/omap: wait for vsync in setcrtc"
This reverts commit 78e65fcba3cf620091c495f12aa62f03b5de02f2.

The purpose of the original patch was to make setcrtc() a blocking
call and make the driver wait till completion of the setcrtc()
operation. However, waiting inside the mode_set() function doesn't
serve the purpose as there's also commit() operation following the
mode_set(). And the driver must be taking steps to wait for
completion of the events inside the commit()->set_enabled() function
rather than inside the mode_set().

Change-Id: Ifca4034a096c8afa533643a282fa5dce5de38fb3
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Somnath Mukherjee <somnath@ti.com>