temp:drm/omap: waiting for pending work during drm_release
It's possible that there are some pending apply_work at the
time when the file is being closed. We must complete all the
pending apply_work at this point of time. Some of them may
even access the per file, event_list, to update the vblank event.
All such accesses must be completed in the preclose().

In the lastclose(), waiting for default framebuffer to come
into effect.

However this patch is marked as "temp" because it doesnt take
care of concurrent usage of the omapdrm driver by more than
one application. In a scenario, while in the process of
flushing out the crtc related work, it may be possible another
application is queue-ing more work. In such a scenario, the
crtc_flush_out() may timeout. This patch currently doesnt handle
a scenario like that.

Change-Id: Ibafe303074b22151f01ac3819ae5d46cf4b84838
Signed-off-by: Somnath Mukherjee <somnath@ti.com>