gc320:Initialize callback function pointers
bltsville userspace registers callbacks to wait for certain events
from drivers.
Based on the status of the return call value of the driver from the callbackevent
appropriate actions are taken. The possible for values for this return value are
GCERR_NONE: status was okay and kernel callback function will be called to complete
            the task
GCERR_TIMEOUT: The wait timed out and an error is retuned to application
GCERR_xxx: Any other error will be signled to Application.

On the driver side, the callback function pointer need to be appripriately updated.
The callback function pointer was not updated in the case of error scenario while
the triggered event ended up with an empty list of commands. This resulted in
uninitialised stack memory being sent to userspace resulting in SIGSEGV

This patch updates the stack values approriately for errorneous scenario.

Change-Id: I721d565d0c59787ca4414a8d0d65257f357c5e3d
Signed-off-by: Pradeep Venkatasubbarao <pradeepv@ti.com>