v4l: ti-vps: vpdma: Add abort channel desc and cleanup APIs
Whenever VPDMA processes a data descriptor of a list, it processes it
and sets up the channel for the DMA transaction. List manager holds the
descriptor in the list till the DMA is complete. If sync_on_channel
descriptor, or another descriptor for same channel are present in the FIFO,
list manager keeps them till the current channel is free.

When the capture stream is closed suddenly while there are pending descriptors
in the FIFO (streamON failed, application killed), it would keep the VPDMA
in busy state. Any further list post would fail with EBUSY.

To avoid this, drivers need to stop the current processing list and cleanup all
the resources VPDMA has taken and also clear the internal FSM of list manager.
The state machiene is cleared by issueing channel specific abort descriptor.

Therefore, the vpdma_list_cleanup accepts an array of channels for which
abort_channel descriptors should be posted. It is driver's decision to post for
all the channels or the channels which were used in the last context.

Change-Id: I64035cb48dab1e48ed682951ba1aca5b4561db45
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>