v4l: ti-vps: vip: Invert clock via SMA_SW1 for falling edge
If the video port uses the pixel clock which is active at falling edge,
then it needs to be inverted inside control module so that, when the
clock reaches the VIP module, the data is valid at rising edge.

This is done by setting bits in the CTRL_CORE_SMA_SW1. This register is
accessed via syscon regmap. Parse the property "syscin-smasw"
from device tree and save the regmap. When configuring the pixel clock
polarity, set the clock inversion bits depending on the instance.

Changed the device tree private data from string to an integer.
This is easier to compare v/s comparing strings.

Change-Id: I26669839a1ff050c8936c90462c34c5a7fd3e4af
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>