drivers: net: cpsw: re-enable tx and rx interrupts separately on open.
When the network interface is brought down using ifconfig, the driver
disables the napi for tx and rx processing separately. Concurrently
if there are Rx packets received or Tx packets completed, we could be
interrupted and the corresponding isr invoked. The ISR checks for the
flag and disables any further interrupts - if a NAPI handler had run it would
have re-enabled the interrupts after servicing completed packets. In the
teardown case, we need to re-check the interrupt enabled flag for rx and
tx separately and re-enable them if need be on subsequent open

Without this fix,  a simple case of concurrent flood ping or iperf traffic
while the interface is brought down can cause the reception/transmission
failure on subsequent device open(as the interrupts were left turned off)

Change-Id: Ib7d886ef581edfd4fcbc2752c8fd6874bdfc9ceb
Signed-off-by: Sriramakrishnan A G <srk@ti.com>