v4l: ti-vps: vpdma: Make list post atomic operation
Writing to the "VPDMA list attribute" register is considered as a list post.
This informs the VPDMA firmware to load the list from the address which
should be taken from the "VPDMA list address" register.

As these two register writes are dependant, it is important that the two writes
happen in atomic manner. This ensures multiple slices (which share same VPDMA)
can post lists asynchronously and all of them point to the correct addresses.

Change-Id: Id7b99b347cd8514ab670f968d49cafb72148c564
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>