i2c: ov1063x: Correct vin6_sel_s0 request check
vin6_sel_s0 is a gpio which is selects between Vin6a and mcasp3
signals. This is needed only when the sensor is to be used is interfaced
via Vin6a (LVDS cam6)

This gpio should not be driven for other sensors/cameras.
Fix a check in the set_gpios function to not drive the gpio
if it is don't care

Change-Id: Ibb4c02f19362ce04727832411b63e36ab604da29
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>