i2c: tvp5158: Configure for line interleaved 4channel muxing
tvp5158 supports multiplexing of upto 4 different NTSC cameras.
This can be used to interface upto 4 cameras using one video port.
Add support for configuring all four decoders simultaneously and
configure the TVP5158 to generate line interleaved format.

As this is embedded sync, there is no way to synchronize the captured
frames from TVP. Also, the order of the lines is not guaranteed to
remain constant.

Change-Id: I630232d2989b76fd664abbfd878b581ab4592e9f
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>