ARM: OMAP: remoteproc: correct watchdog timer data for OMAP4/5 IPU
The watchdog timers chosen for the IPU remoteproc device on OMAP4
and OMAP5 is wrong. The correct watchdog timer for the second
Cortex-M3/M4 core in the IPU is GPTimer 11, and not GPTimer 4.
The timer data was erroneously copied from the corresponding DRA7
watchdog data. GPTimer 4 is used as the SYS/BIOS tick source for
the second core if the IPUs were not running SMP-mode SYS/BIOS.

Signed-off-by: Suman Anna <s-anna@ti.com>