rpmsg: rpc: fix unbalanced mutex locking during recovery
An user thread can be blocked while waiting for new response
messages. The thread is unblocked either when a new message is
available for the thread to read or when the remote processor
has crashed and the rpmsg-rpc device has been removed and
recreated as part of the recovery process.

The user thread currently fails to release a mutex lock it
held in the case of wakeup due to the latter event causing
the user process to never be closed properly or be killed.
Fix this unbalanced mutex locking so that the user process
can proceed to close or recover itself gracefully.

Signed-off-by: Suman Anna <s-anna@ti.com>