rpmsg: rpc: revise couple of traces
Couple of traces have been revised to fix some typos. The
traces in the driver's .remove() function are also switched
to use the device pointer of the rpmsg rpc device being
removed rather than the underlying device pointer of the
representative service's rppc_device structure. This is
done to match the corresponding traces from .probe().

Signed-off-by: Suman Anna <s-anna@ti.com>