omapdss/pll: getting the dss pll supply regulator only once
The vdda_video is the analog supply to all the DSS PLLs. The vdda_video
is supplied by a separate regulator. If we keep asking for the same
regulator using devm_regulator_get(), it will generate a warning message.
This should be ideally done only once as all the PLLs are supplied
by the same regulator. But doing this in omap_dsshw_probe() didn't
work out because due to change in probe order, the regulator framework
couldn't seem to return the regulator corresponding to vdda_video.
Also, the probe deferral is not possible from omap_dsshw_probe().
Hence, the option is to ask for the regulator only once from
dss_dpll_init_regulator() itself.

Change-Id: Iebe43c9a63a371eed1ed55091dfbae89c936874f
Signed-off-by: Somnath Mukherjee <somnath@ti.com>