net/rpmsg: fix memory leak of vrp_channels radix tree
The rpmsg_proto driver is used to provide a socket interface
to userspace under the AF_RPMSG address family, and is used
by the TI IPC MessageQ stack. The usage relies on leveraging
the socket API's connect for Tx sockets and bind for Rx sockets
to exchange messages to/from a remote processor. The MessageQ
stack relies on the Tx sockets leveraging a single rpmsg_channel
device published from the remote processor firmware. All the
published devices from a remote processor are stored in a
corresponding vrp_channels radix tree node, but this is not
freed at all, causing a memory leak.

Fix the memory leak by deleting the radix tree element and
the memory associated with the radix tree node during the
driver's remove.

NOTE:
The MessageQ design relies on only a single published rpmsg
channel device, and as such, this fix deletes the tree node
when that device is unregistered. This needs to be fixed if
multiple devices were to be published from the firmware image.

Signed-off-by: Suman Anna <s-anna@ti.com>